Add all_versions test as a special target.
The test target will attempt to use prove to run the tests in parallel,
falling back to make test if necessary.

Also, I added the -silent option, turned on by default, to keep quiet about
all the other work going on before the tests are run.

Finally, =perl in the command will be substituted with  the current perl
version, and =v will be substituted with the current version.