Be more careful deleting DB structure files.
Retain structure files in the DB if we can't get a new digest for them.

Previously the "digests differ" check would delete structure files if they
were for a non-absolute filename, which couldn't currently be read. This
triggered "interesting" races which could cause the loss of coverage reports
when 2 (or more) programs exited at the same time.