Add digests to DB.
This should fix some problems related to losing coverage data when there are
duplicate files.  This happens most usually when modules are sometimes loaded
from lib and sometimes from blib.

The method is to store digests for each file encountered and to use the first
filename encountered for each digest as the canonical filename.

Using this method, the logic to merge coverage data subsequently should be
able to be deleted.