Fix up problems storing digests with Storable.
The problem was that Storable could eval code which loaded a module, causing
recursion within Devel::Cover and making the import fail.  Make sure we don't
get recursion in this part of Devel::Cover.

Also, move both IO backends into their own modules.  This wasn't a part of the
solution, but it's something I should have done at first, and I did it to
simplfy the logic whilst tracking down this problem.