Manage Devel::Cover coverage reporting errors.
Devel::Cover errors in _report() would end the process.  That wouldn't
normally matter since Devel::Cover tries very hard to be the last thing to be
run anyway.

But it is important just before an exec().  When Devel::Cover sees an exec()
call it first calls report() then calls the exec().  In this case it is
important that any errors in report() do not prevent the exec() from running.

We go belt and braces here: put an eval {} around the _report() and try to
nicely report the problem, and add G_EVAL to call_pv().

Also add a test for this.

This problem was found by Daisuke Maki.  He also diagnosed it, provided the
test case and suggested the solution.