Move closer to using perlbrew and cpanm.
Unfortunately we can't get all the way there yet.  Perlbrew isn't building
early versions of perl that we need yet.  perl-5.6.[12] and 5.8.0 at least
weren't building cleanly for me and, although it doesn't seem to advertise it,
cpanm seems to have a minimum version requirement somewhat higher then this
too.

When we eventually drop support for 5.6.x and, at least, early 5.8.x we can
probably come back to this.  In the meantime we'll comtinue with buildperls.