Move list of ignoired filenames into DB.pm.
Sometimes the main module hasn't been loaded when we want to deal with ignored
filenames, and we don't want to load it because we are not collecting
coverage.  So move the list of ignored filenames into the DB module, since
that will always be loaded when we want to process files.

This fixes https://github.com/pjcj/Devel--Cover/issues/12 and RT 77163.