Get the @INC directories from perl itself.
At build time we put the @INC directories in @Devel::Cover::Inc::Inc so that
we will be able to ignore them by default at runtime.  That works well
normally, but sometimes the build environment differs which can result in
default coverage of core modules.

The solution is to shell out to perl and ask it what its @INC is.  Of course
it's not quite that easy.

When the new V module for introspection is in place we can use that instead.

This is a risky commit.