Experiment to remove dependency on B::Deparse.
Or, at least, for walking the ops to collect coverage.

The experiment seemed to work for statement, subroutine, pod and time
coverage, but didn't for branch and condition coverage.  The main problem is
being able to report on what type of condition was encountered (?:,
if ($cond) { } etc).  If we don't mind losing that reporting capability this
could be a viable approach.

The specific problem is that the context number needs to be passed into
deparse() along with the op to be deparsed and, unless we arrive at this point
from deparse() itself we do not know what that context is.