Provide OpSIBLING if necessary
We also need to provide OpSIBLING, since it is not available on older
perls. This is either an integral part of the patch, if you intend
to push it immediately to begin passing on blead, or a temporary fix,
if you intend to use ppport.h from Devel::PPPort 3.33 or later,
which provides exactly this.