Restore split token used in 'html_basic' report
Fixes #270 

Regression from f6ac419. When the whitespace of Devel::Cover::Reporter::Html_basic was cleaned up, whitespace was added to the token used in the 'split' calls below it. The PPI::HTML document no longer parsed correctly.

Removing the spaces between the word 'class' and the '=' in the $split string fixes the issue.